pyproject.toml
[project]
name = "bookstore-app"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "Evaldoer",email = "evaldorode@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
]

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
